+++ app/src/test/java/io/gnosis/safe/ThreeTenTimezoneDatabaseInit.kt
@@ -1,0 +1,13 @@
+package io.gnosis.safe
+
+import org.threeten.bp.zone.TzdbZoneRulesProvider
+import org.threeten.bp.zone.ZoneRulesProvider
+
+fun Any.initThreeTen() {
+    if (ZoneRulesProvider.getAvailableZoneIds().isEmpty()) {
+        val stream = this.javaClass.classLoader!!.getResourceAsStream("TZDB.dat")
+        stream.use(::TzdbZoneRulesProvider).apply {
+            ZoneRulesProvider.registerProvider(this)
+        }
+    }
+}
